Rename extension from suse-jeos to suse-chost
