+++ .ci/verify
@@ -10,8 +10,8 @@ git config --global user.email "gardener@sap.com"
 git config --global user.name "Gardener CI/CD"
 
 # Required due to https://github.com/kubernetes/kubernetes/issues/86753 - can be removed once the issue is fixed.
-mkdir -p /go/src/github.com/gardener/gardener-extension-os-suse-jeos
-cp -r . /go/src/github.com/gardener/gardener-extension-os-suse-jeos
-cd /go/src/github.com/gardener/gardener-extension-os-suse-jeos
+mkdir -p /go/src/github.com/gardener/gardener-extension-os-suse-chost
+cp -r . /go/src/github.com/gardener/gardener-extension-os-suse-chost
+cd /go/src/github.com/gardener/gardener-extension-os-suse-chost
 
 make verify-extended
