+++ CODEOWNERS
@@ -1,2 +1,2 @@
-# gardener-extension-os-suse-jeos maintainers
-*   @gardener/gardener-extension-os-suse-jeos-maintainers @gardener/gardener-maintainers
+# gardener-extension-os-suse-chost maintainers
+*   @gardener/gardener-extension-os-suse-chost-maintainers @gardener/gardener-maintainers
