+++ charts/gardener-extension-os-suse-chost/Chart.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+appVersion: "1.0"
+description: A Helm chart for the Gardener SUSE CHost extension
+name: gardener-extension-os-suse-chost
+version: 0.1.0
