+++ charts/gardener-extension-os-suse-chost/templates/_helpers.tpl
