+++ charts/gardener-extension-os-suse-chost/templates/deployment.yaml
@@ -1,32 +1,32 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: gardener-extension-os-suse-jeos
+  name: gardener-extension-os-suse-chost
   namespace: {{ .Release.Namespace }}
   labels:
-    app.kubernetes.io/name: gardener-extension-os-suse-jeos
-    helm.sh/chart: gardener-extension-os-suse-jeos
+    app.kubernetes.io/name: gardener-extension-os-suse-chost
+    helm.sh/chart: gardener-extension-os-suse-chost
     app.kubernetes.io/instance: {{ .Release.Name }}
 spec:
   revisionHistoryLimit: 0
   replicas: {{ .Values.replicaCount }}
   selector:
     matchLabels:
-      app.kubernetes.io/name: gardener-extension-os-suse-jeos
+      app.kubernetes.io/name: gardener-extension-os-suse-chost
       app.kubernetes.io/instance: {{ .Release.Name }}
   template:
     metadata:
       labels:
-        app.kubernetes.io/name: gardener-extension-os-suse-jeos
+        app.kubernetes.io/name: gardener-extension-os-suse-chost
         app.kubernetes.io/instance: {{ .Release.Name }}
     spec:
-      serviceAccountName: gardener-extension-os-suse-jeos
+      serviceAccountName: gardener-extension-os-suse-chost
       containers:
-      - name: gardener-extension-os-suse-jeos
+      - name: gardener-extension-os-suse-chost
         image: {{ include "image" . }}
         imagePullPolicy: {{ .Values.image.pullPolicy }}
         command:
-        - /gardener-extension-os-suse-jeos
+        - /gardener-extension-os-suse-chost
         - --max-concurrent-reconciles={{ .Values.controllers.concurrentSyncs }}
         - --disable-controllers={{ .Values.disableControllers | join "," }}
         - --ignore-operation-annotation={{ .Values.controllers.ignoreOperationAnnotation }}
