+++ charts/gardener-extension-os-suse-chost/templates/serviceaccount.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: gardener-extension-os-suse-chost
+  namespace: {{ .Release.Namespace }}
+  labels:
+    app.kubernetes.io/name: gardener-extension-os-suse-chost
+    helm.sh/chart: gardener-extension-os-suse-chost
+    app.kubernetes.io/instance: {{ .Release.Name }}
