--- charts/gardener-extension-os-suse-jeos/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-appVersion: "1.0"
-description: A Helm chart for the Gardener SUSE JeOS extension
-name: gardener-extension-os-suse-jeos
-version: 0.1.0
