--- charts/gardener-extension-os-suse-jeos/templates/serviceaccount.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: gardener-extension-os-suse-jeos
-  namespace: {{ .Release.Namespace }}
-  labels:
-    app.kubernetes.io/name: gardener-extension-os-suse-jeos
-    helm.sh/chart: gardener-extension-os-suse-jeos
-    app.kubernetes.io/instance: {{ .Release.Name }}
