+++ example/controller-registration.yaml
@@ -2,17 +2,19 @@
 apiVersion: core.gardener.cloud/v1beta1
 kind: ControllerRegistration
 metadata:
-  name: os-suse-jeos
+  name: os-suse-chost
 spec:
   resources:
   - kind: OperatingSystemConfig
-    type: suse-jeos
-  - kind: OperatingSystemConfig
     type: suse-chost
+  - kind: OperatingSystemConfig
+    type: memoryone-chost
+  - kind: OperatingSystemConfig
+    type: suse-jeos
   deployment:
     type: helm
     providerConfig:
-      chart: H4sIAAAAAAAAA+0af2/burF/61PcXPShHSrJdurkzcOA+Tlu6y21gzjtQzEMAS3RMluZ1CMpu25f99l3pCRbdtLYSfrStdPBsCT+uDve8Y53J0VEhpRT6dIPmnLFBHeFclWqqPuOCuU/+ApQRzhqtewVYftq7xsHzxrNVvPw0LQ3Wkf1wwfQ+hrEd0GqNJEAD6QQ+rpxu/q/U4h26N+b0njGIi4kvTWNXfpHtW/qv1mvP0P917/iOr8I/+f6fwinRGsquQItIFM0LKaUwzhlcch4BAkJ3pOIKs95COdTpkClSSKkxhvcHDFEsRjDjOhgiqOfgqQx0WxOcZ6eltoJDxEBpxH2Cg6PE0kn7AMNYcFw3J+eeDDk8RIEtzMNS5BQCTHj1HO849HFSCNviKIrZjNE8KY7gpBJ5XgR0779z9h3vPFH6dv/omEa+eaveFRz7q8RjXF9aQITFlPl/NlTiwT/x+Q9/usZ3v8Hh74hkolUQf+4hwQTKd7RQDseCynxs3HY5HhzFYiQ+s631ur+sMv+u1Mitbcks/j2NHbZf/PgaMv/Hx0+q1f2fx9AEvaGSqP3NswbDkmS0qP3s1d3Qzp3QqoCyRJt2zvwEs8ECMzOgImQoKcUXuT7CEavRz34Bx2OYLWjHE5mtA07tpozv4LwtxbPDw+77D8UgReJu9HYYf/NRnM7/jsy3ZX93wP4Ph6DyRJPyqmGx8ETaNYbf4FR5xTQjNG2CbcPZILHIyOaQiBmCeFLDzp49NtpCo98ReWchl4WH5iTFPAaswC3FJ7wKQ9p5iY6GEzgZSQmekEw0jjJhjyFuQdN9BgBTTQQBVxonCdwilwwhdi4nX7S7/YGyJih4Pg+/goMVxBZ4c49GjS9Ojw2A2p5V+3JXw2KpUgxTlkaooAbH3EUi8gZQupm2SgAHtAsXtFrAp7B8TbHIcaa4HCCExJ8mpQHAtE50xamWidt318sFh6xHHtCRn4uNOXna3WR63zWa44RipH2bymTuOLxEtBf4wQyRl5jsrAKiyTFPhPMcVhIDIpM8KVygRs0IVNasnGqN4RW8IhLLw9AseEWqHVG0B/V4JfOqD96apD82j9/OXx9Dr92zs46g/N+bwTDM+gOB8f98/5wgE/PoTN4C//sD46fAmVGkyhODPpwBcgmM+LEHWNwjSjdYKE4U1RCAzZhAS6NRymGoBAJPCS4DUqpnDFl1KpsZIloYjZj2gaX6vK6PAeHRKIdGWdn9rHn+fibUx4K6WPsOE3HHirZLxzi+mblGZU/xVDRL1C4geBaijhG7ylpZIRmqXtqCmUfCl5O7E3vbISyyZ/oB4IyoP6XsJiYC4aJIYULHi2VprOu4BMWtdeor+kPpkJps2qM8G0An5/YlJvtoqAsjDyit5LPG42QjfwCISXGurBmEzbYdJIy9urA/s5g1/mPmyrBfI7epRJ08/rP0VH9WVX/uQ/YX/8XmOyjt1GeTm6YC+7S/zPs26z/NJsHrSr+uw/49MkFCOmEcQyL2AwdeQ3cz58dANPDJjAl6tRWaqCmpqTZOmzXwHtD4pQqz473NIlgNSORjOsJ1B6pvz9S2yMlTYRiWsjldShojDHAFQjbt0bIQ/NQurX3xapDmsRiOaNc5xloJgEMrZSPKXExDdu+tbL+ANjf/tdiumk5aIf9HzQazW37P2pV+d+9QLn+U+z494yHbThe6duZUU1CokkbDWq/Wk42TmFsiIM/fQLvjMaUYBg+KJoz84zJGM3d4AVD3nufjjHAp9oYsvD3pYW5FI1nGHf7Ngjdb8plcozjXuBXcWyYNdmIYVTSOTMYX2IMjI7nxGQdbajbHpuMqWx+7o/yxq5Iuc7WrBBxgFOzVdsC+UlJDHcTxM3XBVDYeM5QSdkG4g3e7srdbfgDKGRv76mcY07XCQIj0sENaJsUBpNz3O4FLnfv7ZyBPVwsm4wHcRquz0yvYHU17DSN41OBul9u7IfsfEpWneV5mH/OMJldy9oFf1/WXHDdGflgktIgxZSNa0wnzYN5rfG3EgPrRM7e54NHSx6oMi8GX8iUyRVLie4Gpry7u+6F3+GdYBxqT2vbuLJ3L67IElZcBeFcZOn6l7jLpgyLGZ3VhDJuyudleWX6POl1jntnF72TXtdUIy4GnVe90Wmn21uNBJgbgs+lmLVLjQATRuPwjE42W/P2U6Kn7ZWFeCsfd4mB4eiiOxw877+4eD48e9U536a7sSWEyvJ2D9NvdAebksvw/TIcniPGV686g+P9cI3xvOrOVoEPRnJrj6REKgO6oe5VYxuHgxZvTfXh8ozfgePhgPsFGvWtmOpbn2QV3Ab2j//kmAS3exG4q/7fqm/nf42Wyf+r+O+PB9d1nXIMaLVMUj0Vkn3MKnzvf7an9Cow7MYoMyrPRExvFRl+RzGfTGPjEV2cyF5IkSaWbXf9alN5BVkviEUaOhue1AwNMmkp+yCKcq2y5drAuurrunzkTqdmBOam4xxlRLW9xhiB2puFiSHtXbK6SxPUC73Meq12mUdFA0m1Kt/7mBuTmH3MON+POE40JEvU1yztxUe26BlJMlboHA+aLfI5jZujK/psApIvu9gbLmo9NDvHBObmhfWVK15sy7a0ujtZ0S/YgIr/4Y0Jl5oHVoXyrpEUjrrsbvaVi0rH5gMda7oZltFG1vCV89hv7cTvAPuf/3neRTIJ3iQS2F3/bWyd/weHjar+cy+w9f3PlcZSlX++bxu/Dva3/3lCbvkd4E77P9z+/rdxcHhU2f99wHbkUsOzWKgAQz8erWOWMXqAZi33DjhWMxxxKsJOPpjKmzsJFzfUno6iKP+hpjAYWxVnrqpdm/ZL9WvTuK8zyWK7vHBnZ2Ytr0SI82tmxbUfyBnssv95VgG60wfAu+y/1Tra/v6/UT+o7P8+IKtoZy8v8lepbaCpFwXSmP5qe+SfvF/5VdSWBWkStcGeGcbyklIZvD8ZCH1qvhZEo3TW2Rp8+uw4ReXSMJMVQtuQfXUMD8Em1y7jzKDMS5tojf6YcX9M1BTMu+GH2JAqaRvXE3AuJRx++gmu6nRd+6HfIwXmseY4pQK04WSrQt6GFjZ+sTDdhgmJFXWcy6XxNvzr384P5DgqqKCCCiqooIIKKqigggoqqKCCCiqooIIKKqigggoq+J+E/wJafkP3AFAAAA==
+      chart: H4sIAAAAAAAAA+0a7W7bOLK/9RRzLrpoD5UUu/nY8+GAcx23DS61gzjtojgcAlqiZbYSqSUpu263++w7pCRbdtLaTnrptqtBEEvkcDic4QxnhoqIDCmn0qXvNeWKCe4K5apMUTeYCKX9e7eHPYSjgwP7i7D+a5+bT/abrYPW4aFpbx7uHR7eg4OvMPdGyJQmEuCeFEJ/CW9T/3cK0Sb9exMaJyziQtKbzrFJ/6j2Vf239lrYBHtfc6Gfg7+4/u/DGdGaSq5AC8j1DLMJ5TDKWBwyHkFKgnckospz7sPFhClQWZoKqfEB90YMUSxGkBAdTBD7MUgaE82mFMfpSaWd8BAJcBphr+DwMJV0zN7TEGYM8f72yIMBj+cguB1pWIKUSogZp57jHQ8vhxp5QxJdkSRI4HV3CCGTyvEipn37P2ff8UYfpG//lw2TyDf/ylc15f6S0AjXl6UwZjFVzt89NUvx/4i8w/86weffEfU1kUxkCk6OezhhKsVbGmjHYyElfo6HTY43VYEIqe98a61uDxvtvzshUntzksQ3nmOT/bf2m+v2v9d8Utv/XQBJ2Wsqjd7bMG06JE0rr97P3p4b0qkTUhVIlmrb3oEXeCRAYDYGjIUEPaHwvNhHMHw17EH3BW4dWGwph5OEtmHTXnOm10z9rQX0g8NG+w9F4EXiVnNssP9W8/BoLf47ah7V9n8n4Pt4DKZzPCknGh4Gj6C11/wHDDtngGaMtk24fSFjPB4Z0RQCkaSEzz3o4NFvhyk88hWVUxp6eXxgTlLA35gFuKXwhM94SHM30cFgAn+GYqxnBCON0xzlMUw9aKHDCGiqgSjgQuM4gUPkjCmkxu3w05Nur4+MmRkc38e/ksI1kyxoFx4NWt4ePDQIjaKr8eifhsRcZBinzM2kgBsfaZSLKBjC2c2yUQA8oHm8opcTeIbGm4KGGGmC6AQHpPg2riIC0QXTFiZap23fn81mHrEce0JGfiE05RdrdZHrYtQrjhGKkfavGZO44tEc0F/jADJCXmMyswqLJMU+E8xxmEkMikzwpQqBGzIhU1qyUaZXhFbyiEuvIqDYcAs0OkM4GTbgaWd4MnxsiPxycvFi8OoCfumcn3f6Fye9IQzOoTvoH59cnAz6+PYMOv038J+T/vFjoMxoEsWJQR+uANlkRpy4YwytIaUrLJRnikppwMYswKXxKMMQFCKBRwS3QSmVCVNGrcpGlkgmZgnTNrhUV9flOYgSiXZknJ3Zx57n49+U8lBIH2PHSTbyUMl+6RCXDwvPqPwJhop+ScINBNdSxDF6T0kjIzQ7u6cmsOJDwStme907H6Jwijf6nqAQqP85MibmgkFq5sIVD+dK06Qr+JhF7Qrt6xESmgg5x1j6i1iWzFsqlBEOJgI2zi8OdsrNrlJQlVkR+FsFFY1GF0bMgZASQ2JYLgZWFuOkVer1qf5ngY3nP+6WFBM6dEg3nmPn+g/G/4f7df3nLmAH/V9ito9uRHk63S0X3KT/fexb1T9mhAd1/HcX8PGjCxDSMeMYFrEEPXQD3E+fHADTw8YwIerMVmqgoSakdXDYboD3msQZVZ7F9zSJYDEilYzrMTQeqH8/UOuYkqZCMY0n05dI0BhjgGsItm9MkIfmpfJon8tVhzSNxTyhXBf5Zy4BDK2UjylxOQzbvrWy/g+wg/0v5bRjOWiD/T9ptlpr9v+k+aRZ2/9dQLX+U+74d4yHbTheqNtJqCYh0aSNBrVlKSdHVBj1IfbHj+Cd05gSjMP7ZXNunzEZob0bwmDm995lI4zwqTaWLPytJ8NsisYJRt6+jS+3HHN1QsZxO/DreDbsmoTEsCrplBmSLzC+Rd9zahKPNuzZHpuPqXx84ZKKxq7IuM5XrZBwgEPzddsa+WlFELcUxe4LAyjNvOCoonAD8Qpzt2bvJgwClNK3z1ROMbHrBIERan+XyU2Ggik6bvqSmLv9ps7BnjGWUcaDOAuXR6dXMrtAO8vi+Eyg/ucreyI/ptJFZ3UcpqEJ5rRLcbvgb82bC66bkPcmOQ0yzMm4xqzSvJjrjX9VOFhmava5QB7OeaCqzBh6IVMmGawkvCuUiu7ushd+g7eCcWg8bqzTyu9gXJEnpLgMwrnI0/bPcZcPGZQjOosBVdqUT6sCyzV62usc984ve6e9rqlKXPY7L3vDs063t8AEmJoJn0mRtCuNAGNG4/Ccjldbi/YzoifthZV4C1d3hYHB8LI76D87eX75bHD+snOxPu/KnhAqz8s9zK/RJ6xKLqf3dDC4QIovX3b6x9vRGuG51U0WARBGdEu3pEQmA7qi7kVjG9FBizemCHF1xG/A8ZDA/QLNvbXY6lufaDXsAjvEf3JEghtdBG6q/x801+//m4eH9f3/nYDruk41BrRKJpmeCMk+5KW7dz/bA3oRGHZjlBmV5yKmN4sMv6uYT2axcYYuDmTPpchSy7i7vNxUXjmvF8QiC50VJ2pQg1xgyr6IshKrbCU2sF76S10+cqczg4Hp6aggGVFtf2Om8oeZiSHtU7p4ylJUDb3KeqNxlUdFA0m1qj77mB6TmH3IOd9uchxopqzMvmRpKz7yRSckzVmhUzxj1qYv5tidXNlnU5Bi2YvN4aLaQ7N3TGRu7qyvXfJsXbiV5d3Kkp5iA2r+r2BQuNoirioV+AVhIdZVr7O1aFQ2Mh/qWPvNyQxXEoevnc9+a2++O+xw/hc5F8lFt0MksLn+u79e/zmq7//vBta+/7nWSOryT/KdWvdm2MH+pym52XeAG+3/aP3+r3mw36rt/y5gPWpp4CEsVIBxH4+W8coIPUCrUXgHxNUMMc5E2CmQqbyBk3BxQ23pKMrSH6oKI7FFVea64rVpv1LANo1bO5M8sitqdnZo3vJShEigYdbc+GF8wUb7n+aln9t8ALzJ/g8Or9z/tw7q+987gbyUnd9cFFepbaCZFwXSmP5iexSfvF/7VdS6AWkStcGeGeYtrRTAT8Z9oc/M54JolM4yV4OPnxynLFkabvIKaBvyz47hPtjU2mWcGZJFTRON0R8x7o+ImoC5HL6PDZmStnE5AMdSwuGnn+C6Tte1X/o9UGBeG45TqTwbTtZK4204wMbPVqTbMCaxoo5ztSbehv/+z/lh/EYNNdRQQw011FBDDTXUUEMNNdRQQw011FBDDTXUUEMNf074A9ByxxEAUAAA
       values:
         image:
           tag: v1.8.0-dev
