+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/gardener/gardener-extension-os-suse-jeos
+module github.com/gardener/gardener-extension-os-suse-chost
 
 go 1.14
