+++ hack/update-codegen.sh
@@ -24,17 +24,17 @@ PROJECT_ROOT=$(dirname $0)/..
 
 bash "${PROJECT_ROOT}"/vendor/k8s.io/code-generator/generate-internal-groups.sh \
   deepcopy,defaulter \
-  github.com/gardener/gardener-extension-os-suse-jeos/pkg/client \
-  github.com/gardener/gardener-extension-os-suse-jeos/pkg/apis \
-  github.com/gardener/gardener-extension-os-suse-jeos/pkg/apis \
+  github.com/gardener/gardener-extension-os-suse-chost/pkg/client \
+  github.com/gardener/gardener-extension-os-suse-chost/pkg/apis \
+  github.com/gardener/gardener-extension-os-suse-chost/pkg/apis \
   "memoryonechost:v1alpha1" \
   --go-header-file "${PROJECT_ROOT}/vendor/github.com/gardener/gardener/hack/LICENSE_BOILERPLATE.txt"
 
 bash "${PROJECT_ROOT}"/vendor/k8s.io/code-generator/generate-internal-groups.sh \
   conversion \
-  github.com/gardener/gardener-extension-os-suse-jeos/pkg/client \
-  github.com/gardener/gardener-extension-os-suse-jeos/pkg/apis \
-  github.com/gardener/gardener-extension-os-suse-jeos/pkg/apis \
+  github.com/gardener/gardener-extension-os-suse-chost/pkg/client \
+  github.com/gardener/gardener-extension-os-suse-chost/pkg/apis \
+  github.com/gardener/gardener-extension-os-suse-chost/pkg/apis \
   "memoryonechost:v1alpha1" \
-  --extra-peer-dirs=github.com/gardener/gardener-extension-os-suse-jeos/pkg/apis/memoryonechost,github.com/gardener/gardener-extension-os-suse-jeos/pkg/apis/memoryonechost/v1alpha1,k8s.io/apimachinery/pkg/apis/meta/v1,k8s.io/apimachinery/pkg/conversion,k8s.io/apimachinery/pkg/runtime \
+  --extra-peer-dirs=github.com/gardener/gardener-extension-os-suse-chost/pkg/apis/memoryonechost,github.com/gardener/gardener-extension-os-suse-chost/pkg/apis/memoryonechost/v1alpha1,k8s.io/apimachinery/pkg/apis/meta/v1,k8s.io/apimachinery/pkg/conversion,k8s.io/apimachinery/pkg/runtime \
   --go-header-file "${PROJECT_ROOT}/vendor/github.com/gardener/gardener/hack/LICENSE_BOILERPLATE.txt"
