+++ pkg/generator/templates/cloud-init/suse-chost.template
