+++ pkg/generator/templates/script/suse-chost.template
